Initial commit with scatter plots and sorted plots in pgfplots.
The module `pandas2pgfplots.py` should be moved outside of this
project as it can be used independently on ltlcross_wrapper and
might be interesting to broader public.

The code in `analyzer.py` only prepares the data and tweaks some
parameters to suit our use-cases.

Currently `ResAnalyzer` can only generate the pgfplots code. I
plan to integrate tikzmagic for Jupyter notebooks in the future.